// src/server/enums/roles.ts
export enum Roles {
    President = 1,
    VicePresident = 2,
    Treasurer = 3,
    FinancialSecretary = 4,
    RecordingSecretary = 5,
    ChairmanoftheBoard = 6,
    ImmediatePastPresident = 7,
    Director = 8,
    Archery = 9,
    AssetManager = 10,
    Bulletin = 11,
    Camping = 12,
    Dinner = 13,
    Fishing = 14,
    GameRoom = 15,
    GroundsandMaintenance = 16,
    Hunting = 17,
    InternalAudit = 18,
    Kitchen = 19,
    LakeandBeach = 20,
    Membership = 21,
    Picnic = 22,
    RifleandPistol = 23,
    SOB = 24,
    Trap = 25,
    WebMaster = 26,
    WomensLiaison = 27
}